emptycart testing in carts controller

+++ app/controllers/carts_controller.rb
@@ -41,8 +41,13 @@ class CartsController < ApplicationController
 
 
   def empty_carts
-    current_cart.destroy
-    session[:cart] = nil
+    if current_cart.line_items.present? 
+      # current_cart.destroy
+      # session[:cart] = nil
+
+  current_cart.destroy if current_cart.id == session[:cart_id]
+  session[:cart_id] = nil
+    end
     redirect_to carts_path
   end
